ChatGPT TELEGRAM BOT/requirements.txt
telethon>=1.28
python-dotenv>=0.21
dateparser>=1.1
python-dateutil>=2.8
langchain>=0.1
openai>=1.0